sql: fix show create table after altering col type of col with comment
Previously `SHOW CREATE TABLE` use `sub_id` from `system.comments`
as a column id to find a column. However we actually use `PGAttributeNum`
when creating a column comment. So things will break if a column's id
has changed such as when type is altered.

Release note (sql change): Previously, if a column in a table has a comment,
`SHOW CREATE TABLE` would fail after the column type is changed.
Now this is fixed.

+++ pkg/sql/catalog/tabledesc/structured_test.go
@@ -917,7 +917,7 @@ func TestLogicalColumnID(t *testing.T) {
 		actual := tests[i].desc.Columns[0].GetPGAttributeNum()
 		expected := tests[i].expected
 
-		if expected != actual {
+		if expected != uint32(actual) {
 			t.Fatalf("Expected PGAttributeNum to be %d, got %d.", expected, actual)
 		}
 	}
